+++ grunt/tasks/bump.js
@@ -1,0 +1,7 @@
+/* global module */
+module.exports = {
+	options: {
+		hardTab: true,
+	},
+	files: ['package.json', 'bower.json']
+};