+++ grunt/tasks/clean.js
@@ -1,0 +1,7 @@
+/* global module */
+module.exports = function(grunt, options) {
+	var path = [options.settings.clean || "dist/*"];
+	return {
+		folder: path
+	};
+};