+++ grunt/tasks/githooks.js
@@ -1,0 +1,6 @@
+/* global module */
+module.exports = {
+	all: {
+		'pre-commit': 'jsbeautifier:precommit jshint:release'
+	}
+};