+++ grunt/tasks/rig.js
@@ -1,0 +1,9 @@
+/* global module */
+module.exports = {
+	all: {
+		expand: true,
+		cwd: "src/build/",
+		src: '*.js',
+		dest: 'dist/'
+	}
+};