+++ grunt/tasks/uglify.js
@@ -1,0 +1,10 @@
+/* global module */
+module.exports = {
+	all: {
+		expand: true,
+		cwd: "dist/",
+		src: '*.js',
+		ext: ".min.js",
+		dest: 'dist/'
+	}
+};