+++ grunt/tasks/watch.js
@@ -1,0 +1,7 @@
+/* global module */
+module.exports = {
+	all: {
+		files: ['Gruntfile.js', 'src/*.*', 'test/**/*'],
+		tasks: ['default']
+	}
+};