+++ History.md
@@ -2,4 +2,4 @@
 0.1.0 / 2014-06-17 
 ==================
 
- * Add jsbeautify.json
+ * Run `pjs-settings` to copy grunt tasks, Gruntfile.js, and jshint/jsbeautify to your project.